BACKED/requirement2.txt
blinker==1.6.2
click==8.1.3
cx-Freeze==6.15.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
patchelf==0.17.2.1
PyMySQL==1.0.3
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
